+++ test/test_aatree.cpp
@@ -20,14 +20,14 @@ TESTDEF::RETPAIR test_AA_TREE()
   Shuffle(testnums, TESTNUM, e);
 
   //uint64_t prof=HighPrecisionTimer::OpenProfiler();
-  for(int i = 0; i<TESTNUM; ++i)
+  for(size_t i = 0; i<TESTNUM; ++i)
     aat.Insert(testnums[i]);
   //std::cout << HighPrecisionTimer::CloseProfiler(prof) << std::endl;
 
   Shuffle(testnums, TESTNUM, e);
   //prof=HighPrecisionTimer::OpenProfiler();
-  uint32_t c = 0;
-  for(int i = 0; i<TESTNUM; ++i)
+  size_t c = 0;
+  for(size_t i = 0; i<TESTNUM; ++i)
     c += (aat.Get(testnums[i]) != 0);
   TEST(c == TESTNUM);
   //std::cout << HighPrecisionTimer::CloseProfiler(prof) << std::endl;
@@ -35,7 +35,7 @@ TESTDEF::RETPAIR test_AA_TREE()
   Shuffle(testnums, TESTNUM, e);
   //prof=HighPrecisionTimer::OpenProfiler();
   c = 0;
-  for(int i = 0; i<TESTNUM; ++i)
+  for(size_t i = 0; i<TESTNUM; ++i)
   {
     //if(testnums[i] == 20159)
     //  std::cout << (aat.Get(testnums[i])!=0) << std::endl;
@@ -58,7 +58,7 @@ TESTDEF::RETPAIR test_AA_TREE()
   aat.Clear();
 
   c = 0;
-  for(int i = 0; i<TESTNUM; ++i) // Test that no numbers are in the tree
+  for(size_t i = 0; i<TESTNUM; ++i) // Test that no numbers are in the tree
     c += (aat.Get(testnums[i]) == 0);
   TEST(c == TESTNUM);
